Added autoaim; made logo & help text move independently
Added auto-aim, which is enabled by default and should be disabled on the introscreen->game transition
and re-enabled on the game->infoscreen transition. Autoaim makes all the blimps aim their cannon towards
the middle of the screen, rather than off into a nonsensical direction.

Made logo and help-text move independently, they are now following a sine curve with a slight lag.