Made game preserve players active-state across rounds. NB: read the disclaimer
If a player set themselves to active, and a new game is started, that player will now automatically be set to be active again the next round. This mechanism relies on the idle timer to set the players to the inactive state again after a certain period of time.

Note that this is implemented through a global variable "roundStates", which is NOT reset upon game-reset (and should be the only variable for which this is true.) The idle-timer needs to explicitly reset this variable to fully reset the games state. roundStates is intended to be used to track a small set of information that is supposed to be preserved across game resets, such as score and active-state. This solution is less than ideal (the reset mechanism should really be more controlled and encapsulated), but since the game is probably not going to obtain a lot of additional gamestates, the added maintenance and debugging burden is hopefully not too great.