Switched to only moving blimps into position when game starts; added green arrow indicators
Switched from a system where blimps are moved into their respective position immediately when the player sets themselves active. This cause a lot of visual "unrest" on the introscreen and made certain configurations a little confusing, so this commit switches the introscreen to a system where players that are active are merely highlighted and slightly raised rather than immediately moved into position.

Additionally, a green bouncing arrow was added to indicate to inactive players that they can grab the controller below the arrow and join the game.