Made game quit after a configurable number of games have been played; other small stuff.
The game can now quit after a certain number of games have been played. An external mechanism like systemd is still required to auto-restart the game. This should alleviate some of the memory leaking issues.

Also quieted some of the debug output and made the stats logger tool be able to deal with restarts.